bugfix: Ensure plugin interops with other plugins
Prior to this PR the babel plugin would traverse within the entering of
the Program node. The unintended consquence of this was that if plugins
that ran before injected Invokables into module scope, for example a "t"
helper for translations, this plugin would never see those injected
bindings.

The PR, moves all the logic to the main visitor and explicitly handles
the 3 cases:

- Template Only Components
- ClassExpression Components
- ClassDeclaration Components

Prior to this we relying on reenterant behavior which made the code hard
to follow and debug.