[BUGFIX] Fix Babel transform interactions with other transforms
While working on some quality of life improvements in the repo, I found
some buggy interactions between the Babel transform and other
transforms. Specifically, the `@babel/preset-typescript` transform and
the `@glimmer/babel-plugin-glimmer-env` transform.

In the first case, the TypeScript transform removes all imports bindings
that no longer have references once types are removed, and it does this
_before_ our transform has a chance to run. This results in imported
values being removed, even though they are used in templates.

In the case of the `@glimmer/babel-plugin-glimmer-env`, referencing the
Program path directly causes errors when it goes to _replace_ references
to the `DEBUG` binding.

This PR updates the logic to create an empty path that is added to the
AST and referenced properly, so TS doesn't remove values and
`@glimmer/env` can "replace" it correctly. We also avoid adding this
extra reference to any _type_ bindings, because then they _won't_ be
removed.

The alternative would be to skip the queue by using the `Program` to
iterate the AST, but that seems like it would be more brittle than this.