[INTERNAL] Adds a new example addon and app
This PR adds a new example addon and app that we can use to demonstrate
Ember/GlimmerX interop. They are a plain, newly generated app and addon,
no changes yet. Just wanted to get them in so the diff is easier to read
later.

Also, I wanted to name the addon `addon` but it wouldn't let me.